+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
+## [Unreleased]
+
 ## [0.14.2] - 2025-04-06
 
 ### Fixed
-- [normalize] Fix gain not correctly being applied to samples
+- [normalize] Fix gain not being applied to samples
 
 ## [0.14.1] - 2025-03-27