+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
-## [Unreleased]
+## [0.15.0] - 2025-04-10
 
 ## Added
 - [player] `--max-ram` command-line option to store audio files in RAM instead of disk (#65)
@@ -362,7 +362,8 @@ Initial release of pleezer, a headless streaming player for the Deezer Connect p
 - Debug and trace logging capabilities
 - Command-line interface with various configuration options
 
-[Unreleased]: https://github.com/roderickvd/pleezer/compare/v0.14.2...HEAD
+[Unreleased]: https://github.com/roderickvd/pleezer/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/roderickvd/pleezer/releases/tag/v0.15.0
 [0.14.2]: https://github.com/roderickvd/pleezer/releases/tag/v0.14.2
 [0.14.1]: https://github.com/roderickvd/pleezer/releases/tag/v0.14.1
 [0.14.0]: https://github.com/roderickvd/pleezer/releases/tag/v0.14.0